Move parsedCommandParser
